fix same name definitions for overview copy buttons
