+++ stories/overview.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Source } from '@storybook/addon-docs';
 import { PACKAGE_VERSION } from '@microsoft/mgt';
 import { versionInfo } from '../.storybook/versionInfo';
-import { TableNamer } from '../.storybook/components/TableNamer';
+import { TableNamer, CopyButtonNamer } from '../.storybook/components/ElementNamer';
 
 <Meta title="Overview" />
 
@@ -105,3 +105,4 @@ The Microsoft Graph Toolkit is supported in the following browsers.
 - Check out the Microsoft Graph Toolkit on [GitHub](https://aka.ms/mgt).
 
 <TableNamer names={['Microsoft Graph Toolkit Components', 'Components in preview', 'Microsoft Graph Toolkit Providers', 'Supported browsers']}/>
+<CopyButtonNamer names={['Copy via mgt-loader', 'Copy via Npm']}/>
